+++ Assets/Scripts/UI/NovelTextBox.cs
