+++ UserSettings/EditorUserSettings.asset
@@ -12,10 +12,10 @@ EditorUserSettings:
       value: 0103025e5d0651035f5b5d7741775a4417161e78292b71357d2b446be4b86d68
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 57575153540d080c555f5d2011730d44444f4c7e7e7d7336292b496ae6e36568
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 57575153540d080c555f5d2011730d44444f4c7e7e7d7336292b496ae6e36568
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
